HS.ability.data {OpenMx}	R Documentation

Holzinger and Swineford (1939) Ability data in 301 children from two schools

Description

This classic data set contains of intelligence-test scores from 301 children on 26 distinct tests. The data are also available in the MBESS package.

The tests cover mental speed, memory, mathematical-ability, spatial, and verbal ability as listed below.

Usage

data("HS.ability.data")

Format

A data frame with 301 observations on the following 2 variables.

id

student ID number (int)

Gender

Sex (Factor w/ 2 levels “Female”,“Male”

grade

Grade in school (integer 7 or 8)

agey

Age in years (integer)

agem

Age in months (integer)

school

School attended (Factor w/2 levels “Grant-White” and “Pasteur”)

addition

A speed test (numeric)

code

A speed test (numeric)

counting

A speed test (numeric)

straight

A speed test (numeric)

wordr

A memory subtest

numberr

A memory subtest

figurer

A memory subtest

object

A memory subtest

numberf

A memory subtest

figurew

A memory subtest

deduct

A mathematical subtest

numeric

A mathematical subtest

problemr

A mathematical subtest

series

A mathematical subtest

arithmet

A mathematical subtest

visual

A spatial subtest

cubes

A spatial subtest

paper

A spatial subtest

flags

A spatial subtest

paperrev

A spatial subtest

flagssub

A spatial subtest

general

A verbal subtest

paragrap

A verbal subtest

sentence

A verbal subtest

wordc

A verbal subtest

wordm

A verbal subtest

Details

The data are from children who differ in grade (seventh- and eighth-grade) and are nested in one of two schools (Pasteur and Grant-White). You will see it in use elsewhere, both in R (lavaan, MBESS), and in Joreskog (1969) reporting a cfa on the Grant-White school subject subset).

The last two tests are substitute versions for other tests. paperrev (a paper form board test) can substitute for paper and flagssub for the lozenges test flags.

Source

Holzinger, K., and Swineford, F. (1939).

References

Holzinger, K., and Swineford, F. (1939). A study in factor analysis: The stability of a bifactor solution. Supplementary Educational Monograph, no. 48. Chicago: University of Chicago Press. Joreskog, K. G. (1969). A general approach to confirmatory maximum likelihood factor analysis. Psychometrika, 34, 183-202.

Examples

data(HS.ability.data)
str(HS.ability.data)
levels(HS.ability.data$school)
plot(flags ~ flagssub, data = HS.ability.data)

[Package OpenMx version 2.9.9 Index]