MxAlgebra-class {OpenMx}	R Documentation

MxAlgebra Class

Description

MxAlgebra is an S4 class. An MxAlgebra object is a named entity. New instances of this class can be created using the function mxAlgebra.

Details

The MxAlgebra class has the following slots:

name 	 - 	 The name of the object

formula 	 - 	 The R expression to be evaluated

result 	 - 	 a matrix with the computation result

The ‘name’ slot is the name of the MxAlgebra object. Use of MxAlgebra objects in the mxConstraint function or an objective function requires reference by name.

The ‘formula’ slot is an expression containing the expression to be evaluated. These objects are operated on or related to one another using one or more operations detailed in the mxAlgebra help file.

The ‘result’ slot is used to hold the results of computing the expression in the ‘formula’ slot. If the containing model has not been executed, then the ‘result’ slot will hold a 0 x 0 matrix. Otherwise the slot will store the computed value of the algebra using the final estimates of the free parameters.

Slots may be referenced with the $ symbol. See the documentation for Classes and the examples in the mxAlgebra document for more information.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

mxAlgebra, mxMatrix, MxMatrix

[Package OpenMx version 2.9.9 Index]