MxConstraint-class {OpenMx}	R Documentation

Class "MxConstraint"

Description

MxConstraint is an S4 class. An MxConstraint object is a named entity. New instances of this class can be created using the function mxConstraint().

Details

Slots may be referenced with the $ symbol. See the documentation for Classes and the examples in the mxConstraint document for more information.

Slots

name:

Character string; the name of the object.

formula:

Object of class "MxAlgebraFormula". The MxAlgebra-like expression representing the constraint function.

alg1:

Object of class "MxCharOrNumber". For internal use.

alg2:

Object of class "MxCharOrNumber". For internal use.

relation:

Object of class "MxCharOrNumber". For internal use.

jac:

Object of class "MxCharOrNumber". Identifies the MxAlgebra representing the Jacobian for the constraint function.

linear:

Logical. For internal use.

Methods

$<-

signature(x = "MxConstraint")

$

signature(x = "MxConstraint")

imxDeparse

signature(object = "MxConstraint")

names

signature(x = "MxConstraint")

print

signature(x = "MxConstraint")

show

signature(object = "MxConstraint")

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

mxConstraint() for the function that creates MxConstraint objects.

Examples

showClass("MxConstraint")

[Package OpenMx version 2.10.0 Index]