MxExpectationGREML-class {OpenMx}	R Documentation

Class "MxExpectationGREML"

Description

MxExpectationGREML is a type of expectation class. It contains the necessary elements for specifying a GREML model. For more information, see mxExpectationGREML().

Objects from the Class

Objects can be created by calls of the form mxExpectationGREML(V, yvars, Xvars, addOnes, blockByPheno, staggerZeroes, dataset.is.yX, casesToDropFromV).

Slots

V:

Object of class "MxCharOrNumber". Identifies the MxAlgebra or MxMatrix to serve as the 'V' matrix.

yvars:

Character vector. Each string names a column of the raw dataset, to be used as a phenotypes.

Xvars:

A list of data column names, specifying the covariates to be used with each phenotype.

addOnes:

Logical; pertains to data-handling at runtime.

blockByPheno:

Logical; pertains to data-handling at runtime.

staggerZeroes:

Logical; pertains to data-handling at runtime.

dataset.is.yX:

Logical; pertains to data-handling at runtime.

y:

Object of class "MxData". Its observed slot will contain the phenotype vector, 'y.'

X:

A matrix, to contain the 'X' matrix of covariates.

yXcolnames:

Character vector; used to store the column names of 'y' and 'X.'

casesToDrop:

Integer vector, specifying the rows and columns of the 'V' matrix to be removed at runtime.

b:

A matrix, to contain the vector of regression coefficients calculated at runtime.

bcov:

A matrix, to contain the sampling covariance matrix of the regression coefficients calculated at runtime.

numFixEff:

Integer number of covariates in 'X.'

dims:

Object of class "character".

numStats:

Numeric; number of observed statistics.

dataColumns:

Object of class "numeric".

name:

Object of class "character".

data:

Object of class "MxCharOrNumber".

.runDims:

Object of class "character".

Extends

Class "MxBaseExpectation", directly. Class "MxBaseNamed", by class "MxBaseExpectation", distance 2. Class "MxExpectation", by class "MxBaseExpectation", distance 2.

Methods

No methods defined with class "MxExpectationGREML" in the signature.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

See mxExpectationGREML() for creating MxExpectationGREML objects, and for more information generally concerning GREML analyses, including a complete example. More information about the OpenMx package may be found here.

Examples

showClass("MxExpectationGREML")

[Package OpenMx version 2.10.0 Index]