MxFitFunctionGREML-class {OpenMx}	R Documentation

Class "MxFitFunctionGREML"

Description

MxFitFunctionGREML is the fitfunction class for GREML analyses.

Objects from the Class

Objects can be created by calls of the form mxFitFunctionGREML(dV).

Slots

dV:

Object of class "MxCharOrNumber". Identifies the MxAlgebra or MxMatrix object(s) to serve as the derivatives of 'V' with respect to free parameters.

dVnames:

Vector of character strings; names of the free parameters corresponding to slot dV.

MLfit:

Object of class "numeric", equal to the maximum-likelihood fitfunction value (as opposed to the restricted maximum-likelihood value).

numObs:

Object of class "integer". Number of observations.

aug:

Object of class "MxCharOrNumber". Identifies the MxAlgebra or MxMatrix object used to "augment" the fitfunction value at each function evaluation during optimization.

augGrad:

Object of class "MxCharOrNumber". Identifies the MxAlgebra or MxMatrix object(s) to serve as the first derivatives of aug with respect to free parameters.

augHess:

Object of class "MxCharOrNumber". Identifies the MxAlgebra or MxMatrix object(s) to serve as the second derivatives of aug with respect to free parameters.

info:

Object of class "list".

dependencies:

Object of class "integer".

expectation:

Object of class "integer".

vector:

Object of class "logical".

rowDiagnostics:

Object of class "logical".

result:

Object of class "matrix".

name:

Object of class "character".

Extends

Class "MxBaseFitFunction", directly. Class "MxBaseNamed", by class "MxBaseFitFunction", distance 2. Class "MxFitFunction", by class "MxBaseFitFunction", distance 2.

Methods

No methods defined with class "MxFitFunctionGREML" in the signature.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

See mxFitFunctionGREML() for creating MxFitFunctionGREML objects. See mxExpectationGREML() for creating MxExpectationGREML objects, and for more information generally concerning GREML analyses, including a complete example. More information about the OpenMx package may be found here.

Examples

showClass("MxFitFunctionGREML")

[Package OpenMx version 2.12.1 Index]