as.statusCode {OpenMx}	R Documentation

Convert a numeric or character vector into an optimizer status code factor

Description

0,‘OK’: Optimization succeeded

1,‘OK/green’: Optimization succeeded, but the sequence of iterates has not yet converged (Mx status GREEN). This condition is only detected by NPSOL.

2,‘infeasible linear constraint’: The linear constraints and bounds could not be satisfied. The problem has no feasible solution.

3,‘infeasible non-linear constraint’: The nonlinear constraints and bounds could not be satisfied. The problem may have no feasible solution.

4,‘iteration limit’: Optimization was stopped prematurely because the iteration limit was reached (Mx status BLUE). You might want to rerun: m1 = mxRun(m1) or increase the iteration limit (see mxOption).

5,‘not convex’: The Hessian at the solution does not appear to be convex (Mx status RED). There may be more than one solution to the model. See mxCheckIdentification.

6,‘nonzero gradient’: The model does not satisfy the first-order optimality conditions to the required accuracy, and no improved point for the merit function could be found during the final linesearch (Mx status RED). To search nearby, see mxTryHard.

7,‘bad deriv’: You have provided analytic derivatives. However, your provided derivatives differ too much from numerically approximated derivatives. Double check your math.

9,‘internal error’: An input parameter was invalid. The most likely cause is a bug in the code. Please report occurrences to the OpenMx developers.

10,‘infeasible start’: Starting values were infeasible. Modify the start values for one or more parameters. For instance, set means to their measured value, or set variances and covariances to plausible values. See mxAutoStart and mxTryHard.

Usage

as.statusCode(code)

Arguments

code
a character or numeric vector of optimizer status code

See Also

mxBootstrap summary.MxModel

[Package OpenMx version 2.12.1 Index]