demoTwoFactor {OpenMx}	R Documentation

Demonstration data for a two factor model

Description

Data set used in some of OpenMx's examples.

Usage

data("demoTwoFactor")

Format

A data frame with 500 observations on the following 10 numeric variables.

x1

x2

x3

x4

x5

y1

y2

y3

y4

y5

Details

Variables x1-x5 are typically used as indicators of one latent trait. Variables y1-y5 are typically used as indicators of another latent trait.

Source

Simulated.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

Examples

data(demoTwoFactor)
cov(demoTwoFactor)
cor(demoTwoFactor)

[Package OpenMx version 2.9.9 Index]