diag2vec {OpenMx}	R Documentation

Extract Diagonal of a Matrix

Description

Given an input matrix, diag2vec returns a column vector of the elements along the diagonal.

Usage

diag2vec(x)

Arguments

x
an input matrix.

Details

Similar to the function diag, except that the input argument is always treated as a matrix (i.e., it doesn't have diag()'s functions of returning an Identity matrix from an nrow specification, nor to return a matrix wrapped around a diagonal if provided with a vector). To get vector2matrix functionality, call vec2diag.

See Also

vec2diag

Examples

diag2vec(matrix(1:9, nrow=3))
#      [,1]
# [1,]    1
# [2,]    5
# [3,]    9

diag2vec(matrix(1:12, nrow=3, ncol=4))
#      [,1]
# [1,]    1
# [2,]    5
# [3,]    9

[Package OpenMx version 2.12.1 Index]