genericFitDependencies,MxBaseFitFunction-method {OpenMx}	R Documentation

Add dependencies

Description

If there is an expectation, then the fitfunction should always depend on it. Hence, subclasses that implement this method must ignore the passed-in dependencies and use "dependencies <- callNextMethod()" instead.

Usage

## S4 method for signature 'MxBaseFitFunction'
genericFitDependencies(.Object, flatModel,
  dependencies)

Arguments

.Object
fit function object

flatModel
flat model that lives with .Object

dependencies
accumulated dependency relationships

[Package OpenMx version 2.9.9 Index]