imxAutoOptionValue {OpenMx}	R Documentation

imxAutoOptionValue

Description

Convert "Auto" placeholders in global mxOptions to actual default values.

Usage

imxAutoOptionValue(optionName, optionList = options()$mxOption)

Arguments

optionName
Character string naming the mxOption for which a numeric or integer value is wanted.

optionList
List of options; defaults to list of global mxOptions. imxAutoOptionValue

Details

This is an internal function exported for documentation purposes. Its primary purpose is to convert the on-load value of "Auto"to valid values for mxOptions ‘Gradient step size’, ‘Gradient iterations’, and ‘Function precision’–respectively, 1.0e-7, 1L, and 1e-14.

[Package OpenMx version 2.10.0 Index]