imxGetSlotDisplayNames {OpenMx}	R Documentation

imxGetSlotDisplayNames

Description

Returns a list of display-friendly object slot names This is an internal function exported for those people who know what they are doing.

Usage

imxGetSlotDisplayNames(object, pattern = ".*",
  slotList = slotNames(object), showDots = FALSE, showEmpty = FALSE)

Arguments

object
The object from which to get slot names

pattern
Initial pattern to match (default of '.*' matches any)

slotList
List of slots for which toget display names (default = slotNames(object), i.e., all)

showDots
Include slots whose names start with '.' (default FALSE)

showEmpty
Include slots with length-zero contents (default FALSE)

[Package OpenMx version 2.12.1 Index]