imxJiggle {OpenMx}	R Documentation

imxJiggle

Description

Jiggle parameter values, subject to box constraints. For internal use by mxTryHard(). This is an internal function exported for those people who know what they are doing.

Usage

imxJiggle(params, lbounds, ubounds, dsn, loc, scale)

Arguments

params
Numeric vector of current free parameter values.

lbounds
Numeric vector of lower bounds on parameters.

ubounds
Numeric vector of upper bounds on parameters.

dsn
Character string naming the family of distribution to be used.

loc
Numeric vector of location parameters (medians).

scale
Numeric vector of scale parameters (see source code).

[Package OpenMx version 2.9.9 Index]