imxModelBuilder {OpenMx}	R Documentation

imxModelBuilder

Description

This is an internal function exported for those people who know what they are doing.

Usage

imxModelBuilder(model, lst, name, manifestVars, latentVars, submodels, remove,
  independent)

Arguments

model
model

lst
lst

name
name

manifestVars
manifestVars

latentVars
latentVars

submodels
submodels

remove
remove

independent
independent

Details

TODO: It probably makes sense to split this into separate methods. For example, modelAddVariables and modelRemoveVariables could be their own methods. This would reduce some cut&paste duplication.

[Package OpenMx version 2.9.9 Index]