imxReplaceSlot {OpenMx}	R Documentation

imxReplaceSlot

Description

Checks for and replaces a slot from the object This is an internal function exported for those people who know what they are doing.

Usage

imxReplaceSlot(x, name, value, check = TRUE)

Arguments

x
object

name
the name of the slot

value
replacement value

check
Check replacement value for validity (default TRUE)

[Package OpenMx version 2.10.0 Index]