imxRobustSE {OpenMx}	R Documentation

imxRobustSE

Description

This is an internal function exported for those people who know what they are doing.

Usage

imxRobustSE(model, details = FALSE)

Arguments

model
An OpenMx model object that has been run

details
logical. whether to return the full parameter covariance matrix

Details

This function computes robust standard errors via a sandwich estimator. The "bread" of the sandwich is the numerically computed inverse Hessian of the likelihood function. This is what is typically used for standard errors throughout OpenMx. The "meat" of the sandwich is proportional to the covariance matrix of the numerically computed row derivatives of the likelihood function (i.e. row gradients).

When details=FALSE, only the standard errors are returned.

When details=TRUE, a list with five named elements is returned. Element SE is the vector of standard errors that is also returned when details=FALSE. Element cov is the full robust covariance matrix of the parameter estimates; the square root of the diagonal of cov gives the standard errors. Element bread is the aforementioned "bread"–the naive (non-robust) covariance matrix of the parameter estimates. Element meat is the aforementioned "meat," proportional to the covvariance matrix of the row gradients. Element TIC is the model's Takeuchi Information Criterion, which is a generalization of AIC calculated from the "bread," the "meat," and the loglikelihood at the maximum-likelihood solution.

This function does not work correctly with multigroup models in which the groups themselves contain subgroups. This function also does not correctly handle multilevel data.

[Package OpenMx version 2.10.0 Index]