imxRowGradients {OpenMx}	R Documentation

imxRowGradients

Description

This is an internal function exported for those people who know what they are doing.

Usage

imxRowGradients(model, robustSE = FALSE)

Arguments

model
An OpenMx model object that has been run

robustSE
Logical; are the row gradients being requested to calculate robust standard errors?

Details

This function computes the gradient for each row of data. The returned object is a matrix with the same number of rows as the data, and the same number of columns as there are free parameters.

[Package OpenMx version 2.12.1 Index]