imxSfClient {OpenMx}	R Documentation

imxSfClient

Description

As of snowfall 1.84, the snowfall supervisor process stores an internal state information in a variable named ".sfOption" that is located in the "snowfall" namespace. The snowfall client processes store internal state information in a variable named ".sfOption" that is located in the global namespace.

Usage

imxSfClient()

Details

As long as the previous statement is true, then the current process is a snowfall client if-and-only-if exists(".sfOption").

[Package OpenMx version 2.12.1 Index]