imxUntitledName {OpenMx}	R Documentation

imxUntitledName

Description

This is an internal function exported for those people who know what they are doing.

Usage

imxUntitledName()

Details

Returns a character, the name of the next untitled entity

[Package OpenMx version 2.12.1 Index]