mxComputeCheckpoint {OpenMx}	R Documentation

Log parameters and state to disk or memory

Description

Captures the current state of the backend. When path is set, the state is written to disk in a single row. When toReturn is set, the state is recorded in memory and returned after mxRun.

Usage

mxComputeCheckpoint(what=NULL, ..., path=NULL, append=FALSE, header=TRUE, toReturn=FALSE,
				parameters=TRUE, loopIndices=TRUE,
  fit=TRUE, counters=TRUE, status=TRUE, standardErrors=FALSE)

Arguments

what
a character vector of algebra names to include in each checkpoint

...
Not used. Forces remaining arguments to be specified by name.

path
a character vector of where to write the checkpoint file

append
if FALSE, truncates the checkpoint file upon open. If TRUE, existing data is preserved and checkpoints are appended.

header
whether to write the header that describes the content of each column

toReturn
logical. Whether to store the checkpoint in memory and return it after the model is run

parameters
logical. Whether to include the parameter vector

loopIndices
logical. Whether to include the loop indices

fit
logical. Whether to include the fit value

counters
logical. Whether to include counters (number of evaluations and iterations)

status
logical. Whether to include the status code

standardErrors
logical. Whether to include the standard errors

[Package OpenMx version 2.12.1 Index]