mxComputeConfidenceInterval {OpenMx}	R Documentation

Find likelihood-based confidence intervals

Description

There are various equivalent ways to pose the optimization problems required to estimate confindence intervals. Most accurate solutions are achieved when the problem is posed using non-linear constraints. However, the available optimizers (CSOLNP, SLSQP, and NPSOL) often have difficulty with non-linear constraints.

Usage

mxComputeConfidenceInterval(plan, ..., freeSet = NA_character_,
  verbose = 0L, engine = NULL, fitfunction = "fitfunction",
  tolerance = NA_real_, constraintType = "none")

Arguments

plan
compute plan to optimize the model

...
Not used. Forces remaining arguments to be specified by name.

freeSet
names of matrices containing free variables

verbose
level of debugging output

engine
deprecated

fitfunction
the name of the deviance function

tolerance
deprecated

constraintType
one of c('ineq', 'none')

References

Neale, M. C. & Miller M. B. (1997). The use of likelihood based confidence intervals in genetic models. Behavior Genetics, 27(2), 113-120.

Pek, J. & Wu, H. (2015). Profile likelihood-based confidence intervals and regions for structural equation models. Psychometrica, 80(4), 1123-1145.

Wu, H. & Neale, M. C. (2012). Adjusted confidence intervals for a bounded parameter. Behavior genetics, 42(6), 886-898.

[Package OpenMx version 2.10.0 Index]