mxComputeIterate {OpenMx}	R Documentation

Repeatedly invoke a series of compute objects until change is less than tolerance

Description

One step (typically the last) must compute the fit or maxAbsChange.

Usage

mxComputeIterate(steps, ..., maxIter = 500L, tolerance = 1e-09,
  verbose = 0L, freeSet = NA_character_, maxDuration = as.numeric(NA))

Arguments

steps
a list of compute objects

...
Not used. Forces remaining arguments to be specified by name.

maxIter
the maximum number of iterations

tolerance
iterates until maximum relative change is less than tolerance

verbose
level of debugging output

freeSet
Names of matrices containing free variables.

maxDuration
the maximum amount of time (in seconds) to iterate

[Package OpenMx version 2.9.9 Index]