mxComputeJacobian {OpenMx}	R Documentation

Numerically estimate the Jacobian with respect to free parameters

Description

When algebra names are given, all algebras must belong to the same model.

When expectations are given, the Jacobian is taken with respect to the manifest model. The manifest model excludes any latent variables or processes. For RAM and LISREL models, the manifest model contains only the manifest variables with free means, covariance, and thresholds. Ordinal manifest variables are standardized.

Usage

mxComputeJacobian(freeSet=NA_character_, ..., of = "expectation",
 defvar.row=as.integer(NA), data='data')

Arguments

freeSet
names of matrices containing free variables

...
Not used. Forces remaining arguments to be specified by name.

of
a character vector of expectations or algebra names

defvar.row
A row index. Which row to load for definition variables.

data
From which data to load definition variables.

See Also

omxManifestModelByParameterJacobian, mxGetExpected

[Package OpenMx version 2.12.1 Index]