mxComputeLoadData {OpenMx}	R Documentation

Load columns into an MxData object

Description

The purpose of this compute step is to help quickly perform many similar analyses. For example, if we are given a sample of people with a few million SNPs (single-nucleotide polymorphism) per person then we could fit a separate model for each SNP by iterating over the SNP data.

Usage

mxComputeLoadData(dest, column, method = "CSV", ..., path,
  originalDataIsIndexOne = FALSE, byrow = TRUE, row.names = FALSE,
  col.names = FALSE, verbose = 0L, cacheSize = 100L)

Arguments

dest
the name of the model where the columns will be loaded

column
a character vector. The column names to replace.

method
name of the conduit used to load the columns.

...
Not used. Forces remaining arguments to be specified by name.

path
when method='CSV', the path to the file containing the data

originalDataIsIndexOne
logical. Whether to use the initial data for index 1

byrow
logical. Whether the data columns are stored in rows (TRUE) or columns (FALSE) on disk.

row.names
logical. Whether row names are present.

col.names
logical. Whether column names are present.

verbose
integer. Level of diagnostic output.

cacheSize
integer. How many columns to cacheSize per scan through the data.

Details

The column names given in the column parameter must already exist in the model's MxData object. The data is assumed to be a placeholder and is not used unless originalDataIsIndexOne is set to TRUE.

When byrow is FALSE, it is necessary to read through the whole file on disk to load a single column. To amortize the cost of reading through the file, cacheSize are loaded each pass through the file.

See Also

mxComputeLoadMatrix, mxComputeCheckpoint

[Package OpenMx version 2.12.1 Index]