mxComputeNewtonRaphson {OpenMx}	R Documentation

Optimize parameters using the Newton-Raphson algorithm

Description

This optimizer requires analytic 1st and 2nd derivatives of the fit function. Box constraints are supported. Parameters can approach box constraints but will not leave the feasible region (even by some small epsilon>0). Non-finite fit values are interpreted as soft feasibility constraints. That is, when a non-finite fit is encountered, line search is continued after the step size is multiplied by 10 available by increasing the verbose level.

Usage

mxComputeNewtonRaphson(freeSet = NA_character_, ...,
  fitfunction = "fitfunction", maxIter = 100L, tolerance = 1e-12,
  verbose = 0L)

Arguments

freeSet
names of matrices containing free variables

...
Not used. Forces remaining arguments to be specified by name.

fitfunction
name of the fitfunction (defaults to 'fitfunction')

maxIter
maximum number of iterations

tolerance
optimization is considered converged when the maximum relative change in fit is less than tolerance

verbose
level of debugging output

References

Luenberger, D. G. & Ye, Y. (2008). Linear and nonlinear programming. Springer.

[Package OpenMx version 2.12.1 Index]