mxComputeStandardError {OpenMx}	R Documentation

Compute standard errors given the Hessian or inverse Hessian

Description

The fit is assumed to be in deviance units (-2 log likelihood).

Usage

mxComputeStandardError(freeSet = NA_character_)

Arguments

freeSet
names of matrices containing free variables

[Package OpenMx version 2.9.9 Index]