mxEvaluateOnGrid {OpenMx}	R Documentation

Evaluate an algebra on an abscissa grid and collect column results

Description

This function evaluates an algebra on a grid of points provided in an auxiliary abscissa matrix.

Usage

mxEvaluateOnGrid(algebra, abscissa)

Arguments

algebra
the name of the single column matrix to be evaluated.

abscissa
the name of the abscissa matrix. See details.

Details

The abscissa matrix must be in a specific format. The variables are in the rows. Abscissa row names must match names of free variables. The grid points are in columns. For each point (column), the free variables are set to the given values and the algebra is re-evaluated. The resulting columns are collected as the result.

Value

Returns the collected columns.

Examples

library(OpenMx)

test2 <- mxModel("test2",
	mxMatrix(values=1.1, nrow=1, ncol=1, free=TRUE, name="thang"),
	mxMatrix(nrow=1, ncol=1, labels="abscissa1", free=TRUE, name="currentAbscissa"),
	mxMatrix(values=-2:2, nrow=1, ncol=5, name="abscissa",
		 dimnames=list(c('abscissa1'), NULL)),
	mxAlgebra(rbind(currentAbscissa + thang, currentAbscissa * thang), name="stuff"),
	mxAlgebra(mxEvaluateOnGrid(stuff, abscissa), name="grid"))
	
test2 <- mxRun(test2)
omxCheckCloseEnough(test2$grid$result, matrix(c(-1:3 + .1, -2:2 * 1.1), ncol=5, nrow=2,byrow=TRUE))

[Package OpenMx version 2.12.1 Index]