mxExpectationBA81 {OpenMx}	R Documentation

Create a Bock & Aitkin (1981) expectation

Description

Used in conjunction with mxFitFunctionML, this expectation models ordinal data with a modest number of latent dimensions. Currently, only a multivariate Normal latent distribution is supported. An equal-interval quadrature is used to integrate over the latent distribution. When all items use the graded response model and items are assumed conditionally independent then item factor analysis is equivalent to a factor model.

Usage

mxExpectationBA81(ItemSpec, item = "item", ..., qpoints = 49L,
  qwidth = 6, mean = "mean", cov = "cov", verbose = 0L,
  weightColumn = NA_integer_, EstepItem = NULL,
  debugInternal = FALSE)

Arguments

ItemSpec
a single item model (to replicate) or a list of item models in the same order as the column of ItemParam

item
the name of the mxMatrix holding item parameters with one column for each item model with parameters starting at row 1 and extra rows filled with NA

...
Not used. Forces remaining arguments to be specified by name.

qpoints
number of points to use for equal interval quadrature integration (default 49L)

qwidth
the width of the quadrature as a positive Z score (default 6.0)

mean
the name of the mxMatrix holding the mean vector

cov
the name of the mxMatrix holding the covariance matrix

verbose
the level of runtime diagnostics (default 0L)

weightColumn
the name of the column in the data containing the row weights (DEPRECATED)

EstepItem
a simple matrix of item parameters for the E-step. This option is mainly of use for debugging derivatives.

debugInternal
when enabled, some of the internal tables are returned in $debug. This is mainly of use to developers.

Details

The conditional likelihood of response x[i,j] to item j from person i with item parameters ξ[j] and latent ability θ[i] is

L(x[i]|ξ,θ[i]) = ∏_j Pr(pick=x[i,j] | ξ[j],θ[i]).

Items are assumed to be conditionally independent. That is, the outcome of one item is assumed to not influence another item after controlling for ξ and θ[i].

The unconditional likelihood is obtained by integrating over the latent distribution θ[i],

L(x[i]|ξ) = \int L(x[i]|ξ, θ[i]) L(θ[i]) d θ[i].

With an assumption that examinees are independently and identically distributed, we can sum the individual log likelihoods,

L=∑_i \log L(x[i] | ξ).

Response models Pr(pick=x[i,j] | ξ[j],θ[i]) are not implemented in OpenMx, but are imported from the RPF package. You must pass a list of models obtained from the RPF package in the ‘ItemSpec’ argument. All item models must use the same number of latent factors although some of these factor loadings can be constrained to zero in the item parameter matrix. The ‘item’ matrix contains item parameters with one item per column in the same order at ItemSpec.

The ‘qpoints’ and ‘qwidth’ argument control the fineness and width, respectively, of the equal-interval quadrature grid. The integer ‘qpoints’ is the number of points per dimension. The quadrature extends from negative qwidth to positive qwidth for each dimension. Since the latent distribution defaults to standard Normal, qwidth can be regarded as a value in Z-score units.

The optional ‘mean’ and ‘cov’ arguments permit modeling of the latent distribution in multigroup models (in a single group, the latent distribution must be fixed). A separate latent covariance model is used in combination with mxExpectationBA81. The point mass distribution contained in the quadrature is converted into a multivariate Normal distribution by mxDataDynamic. Typically mxExpectationNormal is used to fit a multivariate Normal model to these data. Some intricate programming is required. Examples are given in the manual. mxExpectationBA81 uses a sample size of N for the covariance matrix. This differs from mxExpectationNormal which uses a sample size of N-1.

The ‘verbose’ argument enables diagnostics that are mainly of interest to developers.

When a two-tier covariance matrix is recognized, this expectation automatically enables analytic dimension reduction (Cai, 2010).

The optional ‘weightColumn’ is superseded by the weight argument in mxData. For data with many repeated response patterns, model evaluation time can be reduced. An easy way to transform your data into this form is to use compressDataFrame. Non-integer weights are supported except for EAPscores.

mxExpectationBA81 requires mxComputeEM. During a typical optimization run, latent abilities are assumed for examinees during the E-step. These examinee scores are implied by the previous iteration's parameter vector. This can be overridden using the ‘EstepItem’ argument. This is mainly of use to developers for checking item parameter derivatives.

Common univariate priors are available from univariatePrior. The standard Normal distribution of the quadrature acts like a prior distribution for difficulty. It is not necessary to impose any additional Bayesian prior on difficulty estimates (Baker & Kim, 2004, p. 196).

Many estimators are available for standard errors. Oakes is recommended (see mxComputeEM). Also available are Supplement EM (mxComputeEM), Richardson extrapolation (mxComputeNumericDeriv), likelihood-based confidence intervals (mxCI), and the covariance of the rowwise gradients.

References

Bock, R. D., & Aitkin, M. (1981). Marginal maximum likelihood estimation of item parameters: Application of an EM algorithm. Psychometrika, 46, 443-459.

Cai, L. (2010). A two-tier full-information item factor analysis model with applications. Psychometrika, 75, 581-612.

Pritikin, J. N., Hunter, M. D., & Boker, S. M. (2015). Modular open-source software for Item Factor Analysis. Educational and Psychological Measurement, 75(3), 458-474

Pritikin, J. N. & Schmidt, K. M. (in press). Model builder for Item Factor Analysis with OpenMx. R Journal.

Seong, T. J. (1990). Sensitivity of marginal maximum likelihood estimation of item and ability parameters to the characteristics of the prior ability distributions. Applied Psychological Measurement, 14(3), 299-311.

See Also

RPF

Examples

library(OpenMx)
library(rpf)

numItems <- 14

# Create item specifications
spec <- list()
for (ix in 1:numItems) { spec[[ix]] <- rpf.grm(outcomes=sample(2:7, 1)) }
names(spec) <- paste("i", 1:numItems, sep="")

# Generate some random "true" parameter values
correct.mat <- mxSimplify2Array(lapply(spec, rpf.rparam))

# Generate some example data
data <- rpf.sample(500, spec, correct.mat)

# Create a matrix of item parameters with starting values
imat <- mxMatrix(name="item",
                 values=mxSimplify2Array(lapply(spec, rpf.rparam)))
rownames(imat)[1] <- 'f1'
imat$free[!is.na(correct.mat)] <- TRUE
imat$values[!imat$free] <- NA

# Create a compute plan
plan <- mxComputeSequence(list(
  mxComputeEM('expectation', 'scores',
              mxComputeNewtonRaphson(), information="oakes1999",
              infoArgs=list(fitfunction='fitfunction')),
  mxComputeHessianQuality(),
  mxComputeStandardError(),
  mxComputeReportDeriv()))

# Build the OpenMx model
grmModel <- mxModel(model="grm1", imat,
                    mxData(observed=data, type="raw"),
                    mxExpectationBA81(ItemSpec=spec),
                    mxFitFunctionML(),
                    plan)

grmModel <- mxRun(grmModel)
summary(grmModel)

[Package OpenMx version 2.12.1 Index]