mxGetExpected {OpenMx}	R Documentation

Extract the component from a model's expectation

Description

This function extracts the expected means, covariance, or thresholds from a model.

Usage

mxGetExpected(model, component, defvar.row=1, subname=model$name)
imxGetExpectationComponent(model, component, defvar.row=1, subname=model$name)

Arguments

model
MxModel object from which to extract the expectation component.

component
Character vector. The name(s) of the component(s) to extract. Recognized names are “covariance”, “means”, and “thresholds”.

defvar.row
A row index. Which row to load for definition variables.

subname
Name of the submodel to evaluate.

Details

The expected means, covariance, or thresholds can be extracted from Normal (mxExpectationNormal), RAM (mxExpectationRAM), and LISREL (mxExpectationLISREL) models. When more than one component is requested, the components will be returned as a list.

If component 'vector' is requested then the non-redundent coefficients of the expected manifest distribution will be returned as a vector.

If component 'standVector' is requested then the same parameter structure as 'vector' is returned, but it is standardized. For Normal expectations the covariances are returned as correlations, the means are returned as zeros, and the thresholds are returned as z-scores. For the thresholds the z-scores are computed by using the model-implied means and variances.

Note that capitalization is ignored for the 'standVector' option, so 'standvector' is also acceptable.

Value

See details.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

Examples

   
#----------
require(OpenMx)
manifests <- paste("x", 1:5, sep="")
latents <- c("G")
factorModel <- mxModel("One Factor",
      type="RAM",
      manifestVars = manifests,
      latentVars = latents,
      mxPath(from=latents, to=manifests),
      mxPath(from=manifests, arrows=2),
      mxPath(from=latents, arrows=2,
            free=FALSE, values=1.0),
      mxPath(from = 'one', to = manifests))

mxGetExpected(factorModel, "covariance")
# oops.  Starting values indicate a zero covariance matrix.
#  Probably should adjust them.

[Package OpenMx version 2.10.0 Index]