mxRestore {OpenMx}	R Documentation

Restore From Checkpoint File

Description

The function loads the last saved state from a checkpoint file.

Usage

mxRestore(model, chkpt.directory = ".", chkpt.prefix = "", line=NULL, strict=FALSE)

Arguments

model
MxModel object to be loaded.

chkpt.directory
character. Directory where the checkpoint file is located.

chkpt.prefix
character. Prefix of the checkpoint file.

line
integer. Which line from the checkpoint file to restore (defaults to the last line)

strict
logical. Require that the checkpoint name and model name match.

Details

In general, the arguments ‘chkpt.directory’ and ‘chkpt.prefix’ should be identical to the mxOption: ‘Checkpoint Directory’ and ‘Checkpoint Prefix’ that were specificed on the model before execution.

Alternatively, the checkpoint file can be manually loaded as a data.frame in R. Use read.table with the options header=TRUE, sep="\t", stringsAsFactors=FALSE, check.names=FALSE.

Value

Returns an MxModel object with free parameters updated to the last saved values. When ‘line’ is provided, the MxModel is updated to the values on that line within the checkpoint file.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

Examples

library(OpenMx)

# Simulate some data

x=rnorm(1000, mean=0, sd=1)
y= 0.5*x + rnorm(1000, mean=0, sd=1)
tmpFrame <- data.frame(x, y)
tmpNames <- names(tmpFrame)

# Create a model that includes an expected covariance matrix,
# an expectation function, a fit function, and an observed covariance matrix

data <- mxData(cov(tmpFrame), type="cov", numObs = 1000)
expCov <- mxMatrix(type="Symm", nrow=2, ncol=2, values=c(.2,.1,.2), free=TRUE, name="expCov")
expFunction <- mxExpectationNormal(covariance="expCov", dimnames=tmpNames)
fitFunction <- mxFitFunctionML()
testModel <- mxModel(model="testModel", expCov, data, expFunction, fitFunction)

#Use mxRun to optimize the free parameters in the expected covariance matrix
modelOut <- mxRun(testModel, checkpoint = TRUE)
modelOut$expCov

#Use mxRestore to load the last checkpoint saved state of the model
modelRestore <- mxRestore(testModel)
modelRestore$expCov

[Package OpenMx version 2.10.0 Index]