mxRetro {OpenMx}	R Documentation

Return random classic Mx error message

Description

This function allows you to obtain a classic Mx error message. The message returned is random.

Usage

mxRetro()

Details

If you're a nostalgic old sod and you miss the warm, fuzzy feelings you got from reading one of Mike Neale's pattented error messages, then this function is here to save you from the depths of dire depression. All credit for this function is due to Sarah Medland, but of course the wisdom is from Mike Neale.

References

- https://en.wikipedia.org/wiki/OpenMx

See Also

- omxBrownie

Examples

require(OpenMx)
mxRetro()

[Package OpenMx version 2.10.0 Index]