mxRun {OpenMx}	R Documentation

Send a Model to the Optimizer

Description

This function begins optimization, sending ‘model’ to the optimizer.

Optionally, confidence ‘intervals’ will also be computed.

During a run, ‘mxRun’ will print the context (e.g. optimizer name or step in the analysis e.g. MxComputeNumericDeriv ), followed by the current evaluation count, fit value, and the change in fit compared to the last status report. e.g.:

SLSQP 668 4966.58 -218.8

note: For models that prove difficult to run, you might look at mxTryHard.

Usage

mxRun(model, ..., intervals = NULL, silent = FALSE, suppressWarnings = FALSE, 
    unsafe = FALSE, checkpoint = FALSE, useSocket = FALSE, onlyFrontend = FALSE, 
    useOptimizer = TRUE)

Arguments

model
A MxModel object to be optimized.

...
Not used. Forces remaining arguments to be specified by name.

intervals
A boolean indicating whether to compute the specified confidence intervals.

silent
A boolean indicating whether to print status to terminal.

suppressWarnings
A boolean indicating whether to suppress warnings.

unsafe
A boolean indicating whether to ignore errors.

checkpoint
A boolean indicating whether to periodically write parameter values to a file.

useSocket
A boolean indicating whether to periodically write parameter values to a socket.

onlyFrontend
A boolean indicating whether to run only front-end model transformations.

useOptimizer
A boolean indicating whether to run only the log-likelihood of the current free parameter values but not move any of the free parameters.

Details

The mxRun function is used to optimize free parameters in MxModel objects based on an expectation function and fit function. MxModel objects included in the mxRun function must include an appropriate expectation and fit functions.

If the ‘silent’ flag is TRUE, then model execution will not print any status messages to the terminal.

If the ‘suppressWarnings’ flag is TRUE, then model execution will not issue a warning if NPSOL returns a non-zero status code.

If the ‘unsafe’ flag is TRUE, then many error conditions will not be detected. Any error condition detected will be downgraded to warnings. It is strongly recommended to use this feature only for debugging purposes.

Free parameters are estimated or updated based on the expectation and fit functions. These estimated values, along with estimation information and model fit, can be found in the 'output' slot of MxModel objects after mxRun has been used.

If a model is dependent on or shares parameters with another model, both models must be included as arguments in another MxModel object. This top-level MxModel object must include expectation and fit functions in both submodels, as well as an additional fit function describing how the results of the first two should be combined (e.g. mxFitFunctionMultigroup).

Value

Returns an MxModel object with free parameters updated to their final values. The return value contains an "output" slot with the results of optimization.

References

The OpenMx User's guide can be found at https://openmx.ssri.psu.edu/documentation.

See Also

mxTryHard for running models which prove difficult to optimize; summary to print a summary of a run model; mxModel for more on the model itself; More information about the OpenMx package may be found here.

Examples

# Create and run the 1-factor CFA on the openmx.ssri.psu.edu front page

# 1. Load OpenMx and the demoOneFactor dataframe

library(OpenMx)
data(demoOneFactor)  

# 2. Define the manifests (5 demo variables) and latents for use in the model

manifests <- names(demoOneFactor) 
latents   <- c("G")

# 3. Build the model, adding paths and data
model <- mxModel(model="One Factor", type="RAM",
    manifestVars = manifests,
    latentVars   = latents,
    mxPath(from=latents, to=manifests, labels=paste("b", 1:5, sep="")),
    mxPath(from=manifests, arrows=2, labels=paste("u", 1:5, sep="")),
    mxPath(from=latents  , arrows=2, free=FALSE, values=1.0),
    mxData(cov(demoOneFactor), type="cov", numObs=500)
)

# 4. Run the model, returning the result into model
model <- mxRun(model) 

# 5. Show a summary of the fitted model and parameter values
summary(model) 

[Package OpenMx version 2.12.1 Index]