mxVersion {OpenMx}	R Documentation

Returns Current Version String

Description

This function returns a string with the current version number of OpenMx. Optionally (with verbose = TRUE (the default)), it prints a message containing the version of R, the platform, and the optimizer.

Usage

mxVersion(model = NULL, verbose = TRUE)

Arguments

model
optional MxModel to request optimizer from (default = NULL)

verbose
Whether to print version information to the console (default = TRUE)

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

Examples

# Print useful version information.
mxVersion()
# If you just want the version, use this call.
x = mxVersion(verbose=FALSE)

library(OpenMx)
data(demoOneFactor)  # load the demoOneFactor dataframe
manifests <- names(demoOneFactor) # set the manifest to the 5 demo variables
latents   <- c("G")  # define 1 latent variable
model <- mxModel(model = "One Factor", type = "RAM",
    manifestVars = manifests,
    latentVars   = latents,
    mxPath(from = latents, to = manifests, labels = paste("b", 1:5, sep = "")), 
    mxPath(from = manifests, arrows = 2  , labels = paste("u", 1:5, sep = "")), 
    mxPath(from = latents  , arrows = 2  , free = FALSE, values = 1.0), 
    mxData(cov(demoOneFactor), type = "cov", numObs = 500)
)
mxVersion(model, verbose = TRUE)

[Package OpenMx version 2.12.1 Index]