myTwinData {OpenMx}	R Documentation

Twin data on weight and height

Description

Data set used in some of OpenMx's examples.

Usage

data("myTwinData")

Format

A data frame with 3808 observations on the following variables.

fam

Family ID variable

age

Age of the twin pair. Range: 17 to 88.

zyg

Integer codes for zygosity and gender combinations

part

wt1

Weight in kilograms for twin 1

wt2

Weight in kilograms for twin 2

ht1

Height in meters for twin 1

ht2

Height in meters for twin 2

htwt1

Product of ht and wt for twin 1

htwt2

Product of ht and wt for twin 2

bmi1

Body Mass Index for twin 1

bmi2

Body Mass Index for twin 2

Details

Height and weight are highly correlated, and each individually highly heritable. These data present and opportunity for multivariate behavior genetics modeling.

Source

Timothy Bates

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

Examples

data(myTwinData)

plot( ht1 ~ wt1, myTwinData)

[Package OpenMx version 2.9.9 Index]