omxBrownie {OpenMx}	R Documentation

Make Brownies in OpenMx

Description

This function returns a brownie recipe.

Usage

omxBrownie(quantity=1, walnuts=TRUE)

Arguments

quantity
Number of batches of brownies desired. Defaults to one.

walnuts
Logical. Indicates whether walnuts are to be included in the brownies. Defaults to TRUE.

Details

Returns a brownie recipe. Alter the 'quantity' variable to make more pans of brownies. Ingredients, equipment and procedure are listed, but neither ingredients nor equipment are provided.

Raw cocoa powder can be used instead of Dutch processed cocoa for approximately double the antioxidants and flavonols. However, raw cocoa powder is not as smooth and delicious in taste.

We substitute coconut butter for dairy butter because dairy butter contains a large proportion of saturated fat that raises deadly LDL cholesterol (Trumbo & Shimakawa, 2011). In contrast, coconut butter has so much fiber that the considerable saturated fat that it contains is mostly not absorbed (Padmakumaran, Rajamohan & Kurup, 1999). You can substitute erythritol (den Hartog et al, 2010) for sucanat (Lustig, Schmidt, & Brindis, 2012) to improve the glycemic index and calorie density. We substitute whole wheat flour for all-purpose wheat flour because whole grains are associated with improvement in blood pressure (Tighe et al, 2010).

Value

Returns a brownie recipe.

References

Padmakumaran Nair KG, Rajamohan T, Kurup PA. (1999). Coconut kernel protein modifies the effect of coconut oil on serum lipids. Plant Foods Hum Nutr. 53(2):133-44.

Tighe P, Duthie G, Vaughan N, Brittenden J, Simpson WG, Duthie S, Mutch W, Wahle K, Horgan G, Thies F. (2010). Effect of increased consumption of whole-grain foods on blood pressure and other cardiovascular risk markers in healthy middle-aged persons: a randomized controlled trial. Am J Clin Nutr. 92(4), 733-40.

R H Lustig, L A Schmidt, C D Brindis. (2012). Public health: The toxic truth about sugar. Nature. 482 27-29.

den Hartog GJ, Boots AW, Adam-Perrot A, Brouns F, Verkooijen IW, Weseler AR, Haenen GR, Bast A. (2010). Erythritol is a sweet antioxidant. Nutrition. 26(4), 449-58.

Trumbo PR, Shimakawa T. (2011). Tolerable upper intake levels for trans fat, saturated fat, and cholesterol. Nutr Rev. 69(5), 270-8. doi: 10.1111/j.1753-4887.2011.00389.x.

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

More information about the OpenMx package may be found here.

Examples

# Return a brownie recipe
omxBrownie()

[Package OpenMx version 2.9.9 Index]