omxCheckEquals {OpenMx}	R Documentation

Equality Testing Function

Description

This function tests whether two objects are equal using the ‘==’ operator.

Usage

omxCheckEquals(a, b)

Arguments

a
the first value to compare.

b
the second value to compare.

Details

Performs the ‘==’ comparison on the two arguments. If the two arguments are not equal, then an error will be thrown. If ‘a’ and ‘b’ are equal to each other, by default the function will print a statement informing the user the test has passed. To turn off these print statements use options("mxPrintUnitTests" = FALSE).

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

omxCheckCloseEnough, omxCheckWithinPercentError, omxCheckSetEquals, omxCheckTrue, omxCheckIdentical

Examples

omxCheckEquals(c(1, 2, 3), c(1, 2, 3))

omxCheckEquals(FALSE, FALSE)

# Throws an error
try(omxCheckEquals(c(1, 2, 3), c(2, 1, 3)))

[Package OpenMx version 2.10.0 Index]