omxCheckError {OpenMx}	R Documentation

Correct Error Message Function

Description

This function tests whether the correct error message is thrown.

Usage

omxCheckError(expression, message)

Arguments

expression
an R expression that produces an error

message
a character string with the desired error message

Details

Arguments ‘expression’ and ‘message’ give the expression that generates the error and the message that is supposed to be generated, respectively.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

See Also

omxCheckWarning omxCheckWithinPercentError, omxCheckIdentical, omxCheckSetEquals, omxCheckTrue, omxCheckEquals

Examples

A <- mxMatrix('Full', 1, 1, labels = 'data.foo', free = TRUE, name = 'A')
model <- mxModel('model', A)
omxCheckError(mxRun(model), 
paste("The definition variable 'data.foo'",
	"has been assigned to a",
	"free parameter in matrix 'A'"))
omxCheckCloseEnough(matrix(3, 3, 3), matrix(4, 3, 3), epsilon = 2)
# Throws error, check the message
tmsg <- paste("In omxCheckCloseEnough(c(1, 2, 3), c(1.1, 1.9, 3), 0.01)",
	": not equal to within 0.01 : '1 2 3' and '1.1 1.9 3'")
omxCheckError(omxCheckCloseEnough(c(1, 2, 3), c(1.1, 1.9 ,3.0), .01), tmsg)

[Package OpenMx version 2.9.9 Index]