omxConstrainMLThresholds {OpenMx}	R Documentation

omxConstrainMLThresholds

Description

Add constraint to ML model to keep thresholds in order

Usage

omxConstrainMLThresholds(model, dist = 0.1)

Arguments

model
the MxModel to which constraints should be added

dist
unused

Details

This function adds a nonlinear constraint to an ML model. The constraint keeps the thresholds in order. Constraints often slow model estimation, however, keeping the thresholds in increasing order helps ensure the likelihood function is well-defined. If you're having problems with ordinal data, this is one of the things to try.

Value

a new MxModel object with the constraints added

See Also

demo("omxConstrainMLThresholds")

[Package OpenMx version 2.12.1 Index]