omxDefaultComputePlan {OpenMx}	R Documentation

Construct default compute plan

Description

This function generates a default compute plan, where "default compute plan" refers to an object of class MxComputeSequence which is appropriate for use in a wide variety of cases. The exact specification of the plan will depend upon the arguments provided to omxDefaultComputePlan().

Usage

omxDefaultComputePlan(modelName=NULL, intervals=FALSE,
  useOptimizer=TRUE, optionList=options()$mxOption)

Arguments

modelName
Optional (defaults to NULL) character string, providing the name of the MxModel the fitfunction of which is to be evaluated, and usually, optimized.

intervals
Logical; will confidence intervals be computed? Defaults to FALSE.

useOptimizer
Logical; will a fitfunction be minimized? Defaults to TRUE.

optionList
List of mxOptions. Defaults to the current list of global mxOptions.

Details

At minimum, argument optionList must include “Gradient algorithm”, “Gradient iterations”, “Gradient step size”, “Calculate Hessian”, and “Standard Errors”. Users are advised to set “Calculate Hessian” and “Standard Errors” to "No" if the compute plan is to be used in an MxModel containing MxConstraints.

Value

Returns an object of class MxComputeSequence.

Examples

foo <- omxDefaultComputePlan(modelName="bar")
str(foo)

[Package OpenMx version 2.10.0 Index]