omxGetBootstrapReplications {OpenMx}	R Documentation

omxGetBootstrapReplications

Description

Checks a variety of conditions to ensure that bootstrap replications are available and valid. Throws exception if things go wrong. Otherwise, replications are returned to the caller.

Usage

omxGetBootstrapReplications(model)
omxBootstrapCov(model)

Arguments

model
an MxModel object

Value

a matrix or covariance matrix of bootstrap parameter estimates

[Package OpenMx version 2.9.9 Index]