omxGetRAMDepth {OpenMx}	R Documentation

omxGetRAMDepth

Description

Get the potency of a matrix for inversion speed-up

Usage

omxGetRAMDepth(A, maxdepth = nrow(A) - 1)

Arguments

A
MxMatrix object

maxdepth
Numeric. maximum depth to check

Details

This function is used internally by the mxExpectationRAM function to determine how far to expand (I-A)^{-1} = I + A + A^2 + A^3 + .... It is similarly used by mxExpectationLISREL in expanding (I-B)^{-1} = I + B + B^2 + B^3 + .... In many situations A^2 is a zero matrix (nilpotent of order 2). So when A has large dimension it is much faster to compute I+A than (I-A)^{-1}.

[Package OpenMx version 2.9.9 Index]