omxGraphviz {OpenMx}	R Documentation

Show RAM Model in Graphviz Format

Description

The function accepts a RAM style model and outputs a visual representation of the model in Graphviz format. The function will output either to a file or to the console. The recommended file extension for an output file is ".dot".

Usage

omxGraphviz(model, dotFilename = "")

Arguments

model
An RAM-type model.

dotFilename
The name of the output file. Use "" to write to console.

Value

Invisibly returns a string containing the model description in Graphviz format.

References

The OpenMx User's guide can be found at http://openmx.ssri.psu.edu/documentation.

[Package OpenMx version 2.12.1 Index]