omxLapply {OpenMx}	R Documentation

On-Demand Parallel Lapply

Description

If the snowfall library is loaded, then this function calls sfLapply. Otherwise it invokes lapply.

Usage

omxLapply(x, fun, ...)

Arguments

x
a vector (atomic or list) or an expressions vector. Other objects (including classed objects) will be coerced by as.list.

fun
the function to be applied to each element of x.

...
optional arguments to fun.

See Also

omxApply, omxSapply

Examples

x <- list(a = 1:10, beta = exp(-3:3), logic = c(TRUE,FALSE,FALSE,TRUE))
# compute the list mean for each list element
omxLapply(x,mean)

[Package OpenMx version 2.12.1 Index]