omxLocateParameters {OpenMx}	R Documentation

Summarize Model Parameters

Description

Return a data.frame object summarizing the free parameters in the model.

Usage

omxLocateParameters(model, labels = NULL, indep = FALSE)

Arguments

model
a MxModel object

labels
optionally specify which free parameters to retrieve.

indep
fetch parameters from independent submodels.

Details

Invoking the function with the default value for the ‘labels’ argument retrieves all the free parameters. The ‘labels’ argument can be used to select a subset of the free parameters. Note that ‘NA’ is a valid possible input to the ‘labels’ argument.

See Also

omxGetParameters, omxSetParameters, omxAssignFirstParameters

Examples

A <- mxMatrix('Full', 2, 2, labels = c("A11", "A12", NA, NA), values= 1:4,
   free = TRUE, byrow = TRUE, name = 'A')

model <- mxModel(A, name = 'model')

# Request all free parameters in model
omxLocateParameters(model)

# Request free parameters "A11" and all NAs
omxLocateParameters(model, c("A11", NA))

[Package OpenMx version 2.10.0 Index]