omxMatrixOperations {OpenMx}	R Documentation

MxMatrix operations

Description

omxCbind columnwise binding of two or more MxMatrices. omxRbind rowwise binding of two or more MxMatrices. omxTranspose transpose of MxMatrix.

Usage

omxCbind(..., allowUnlabeled = 
    getOption("mxOptions")[["Allow Unlabeled"]], 
    dimnames = NA, name = NA)
omxRbind(..., allowUnlabeled = 
    getOption("mxOptions")[["Allow Unlabeled"]], 
    dimnames = NA, name = NA)
omxTranspose(matrix, allowUnlabeled = 
    getOption("mxOptions")[["Allow Unlabeled"]], 
    dimnames = NA, name = NA)

Arguments

...
two or more MxMatrix objects

matrix
MxMatrix input

allowUnlabeled
whether or not to accept free parameters with NA labels

dimnames
list. The dimnames attribute for the matrix: a list of length 2 giving the row and column names respectively. An empty list is treated as NULL, and a list of length one as row names. The list can be named, and the list names will be used as names for the dimensions.

name
an optional character string indicating the name of the MxMatrix object

[Package OpenMx version 2.10.0 Index]