omxRMSEA {OpenMx}	R Documentation

Get the RMSEA with confidence intervals from model

Description

This function calculates the Root Mean Square Error of the Approximation (RMSEA) for a model and computes confidence intervals for that fit statistic.

Usage

omxRMSEA(model, lower=.025, upper=.975, null=.05, ...)

Arguments

model
An MxModel object for which the RMSEA is desired

lower
The lower confidence bound for the confidence interval

upper
The upper confidence bound for the confidence interval

null
Value of RMSEA used to test for close fit

...
Further named arguments passed to summary

Details

To help users obtain fit statistics related to the RMSEA, this function confidence intervals and a test for close fit. The user determines how close the fit is required to be by setting the null argument to the value desired for comparison.

Value

A named vector with elements lower, est.rmsea, upper, null, and 'Prob(x <= null)'.

References

Browne, M. W. & Cudeck, R. (1992). Alternative Ways of Assessing Model Fit. Sociological Methods and Research, 21, 230-258.

Examples

require(OpenMx)
data(demoOneFactor)
manifests <- names(demoOneFactor)
latents <- c("G")
factorModel <- mxModel("One Factor", 
                       type="RAM",
                       manifestVars=manifests, 
                       latentVars=latents,
                       mxPath(from=latents, to=manifests),
                       mxPath(from=manifests, arrows=2),
                       mxPath(from=latents, arrows=2, free=FALSE, values=1.0),
                       mxData(observed=cov(demoOneFactor), type="cov", numObs=500))
factorRun <- mxRun(factorModel)
factorSat <- mxRefModels(factorRun, run=TRUE)
summary(factorRun, refModels=factorSat)
# Gives RMSEA with 95% confidence interval

omxRMSEA(factorRun, .05, .95, refModels=factorSat)
# Gives RMSEA with 90% confidence interval
#  and probability of 'close enough' fit

[Package OpenMx version 2.12.1 Index]