omxSapply {OpenMx}	R Documentation

On-Demand Parallel Sapply

Description

If the snowfall library is loaded, then this function calls sfSapply. Otherwise it invokes sapply.

Usage

omxSapply(x, fun, ..., simplify = TRUE, USE.NAMES = TRUE)

Arguments

x
a vector (atomic or list) or an expressions vector. Other objects (including classed objects) will be coerced by as.list.

fun
the function to be applied to each element of x.

...
optional arguments to fun.

simplify
logical; should the result be simplified to a vector or matrix if possible?

USE.NAMES
logical; if TRUE and if x is a character, use x as names for the result unless it had names already.

See Also

omxApply, omxLapply

Examples

x <- list(a = 1:10, beta = exp(-3:3), logic = c(TRUE,FALSE,FALSE,TRUE))
# compute the list mean for each list element
omxSapply(x, quantile)

[Package OpenMx version 2.12.1 Index]