omxSetParameters {OpenMx}	R Documentation

Assign Model Parameters

Description

Modify the attributes of parameters in a model. This function cannot modify parameters that have NA labels. Often you will want to call omxAssignFirstParameters after using this, to force the starting values of equated parameters to the same value (otherwise the model cannot begin to be evaluated)

Usage

omxSetParameters(model, labels, free = NULL, values = NULL,
    newlabels = NULL, lbound = NULL, ubound = NULL, indep = FALSE,
    strict = TRUE, name = NULL)

Arguments

model
an MxModel object.

labels
a character vector of target parameter names.

free
a boolean vector of parameter free/fixed designations.

values
a numeric vector of parameter values.

newlabels
a character vector of new parameter names.

lbound
a numeric vector of lower bound values.

ubound
a numeric vector of upper bound values.

indep
boolean. set parameters in independent submodels.

strict
boolean. If TRUE then throw an error when a label does not appear in the model.

name
character string. (optional) a new name for the model.

See Also

omxGetParameters, omxAssignFirstParameters

Examples

A <- mxMatrix('Full', 3, 3, labels = c('a','b', NA), free = TRUE, name = 'A')
model <- mxModel(model="testModel", A, name = 'model')

# set value of cells labelled "a" and "b" to 1 and 2 respectively
model <- omxSetParameters(model, c('a', 'b'), values = c(1, 2))

# set label of cell labelled "a" to "b" and vice versa
model <- omxSetParameters(model, c('a', 'b'), newlabels = c('b', 'a'))

# set label of cells labelled "a" to "b"
model <- omxSetParameters(model, c('a'), newlabels = 'b')

# ensure initial values are the same for each instance of a labeled parameter
model <- omxAssignFirstParameters(model)

[Package OpenMx version 2.12.1 Index]