twinData {OpenMx}	R Documentation

Australian twin sample biometric data.

Description

Australian twin data with 3,808 observations on the 12 variables including body mass index (BMI) assessed in both MZ and DZ twins.

Questionnaires were mailed to 5,967 pairs age 18 years and over. These data consist of completed questionnaires returned by both members of 3,808 (64 percent) pairs. There are two cohort blocks in the data: a younger group (zyg 1:5), and an older group (zyg 6:10)

It is a wide dataset, with two individuals per line. Families are identified by the variable “fam”.

Data include zygosity (zyg), along with heights in metres, weights in kg, and the derived variables BMI in kg/m^2 (stored as “htwt1” and “htwt2”), as well as the 7 times the natural log of this variable, stored as bmi1 and bmi2. The logged values are more closely normally distributed while scaling by 7 places them into a similar range to the original variable.

For convenience, zyg is broken out into separate “zygosity” and “cohort” factors. “zygosity” is coded as a factor with 5-levels: MZFF, MZMM, DZFF, DZMM, DZOS. DZOS are in Female/Male wide order.

Usage

data(twinData)

Format

A data frame with 3808 observations on the following 12 variables.

fam

The family ID

age

Age in years (of both twins)

zyg

Code for zygosity and cohort (see details)

part

A numeric vector

wt1

Weight of twin 1 (kg)

wt2

Weight of twin 2 (kg)

ht1

Height of twin 1 (m)

ht2

Height of twin 2 (m)

htwt1

Raw BMI of twin 1 (kg/m^2)

htwt2

Raw BMI of twin 2 (kg/m^2)

bmi1

7*log(BMI) of twin 1

bmi2

7*log(BMI) of twin 2

cohort

Either “younger” or “older”

zygosity

Zygosity factor with levels: MZFF, MZMM, DZFF, DZMM, DZOS

age1

Age of Twin 1

age2

Age of Twin 2

Details

“zyg” codes twin-zygosity as follows: 1 == MZFF (i.e MZ females) 2 == MZMM (i.e MZ males) 3 == DZFF 4 == DZMM 5 == DZOS opposite sex pairs

Note: zyg 6:10 are for an older cohort in the sample. So: 6 == MZFF (i.e MZ females) 7 == MZMM (i.e MZ males) 8 == DZFF 9 == DZMM 10 == DZOS opposite sex pairs

The “zygosity” and “cohort” variables take care of this for you (conventions differ).

References

Martin, N. G. & Jardine, R. (1986). Eysenck's contribution to behavior genetics. In S. Modgil & C. Modgil (Eds.), Hans Eysenck: Consensus and Controversy. Falmer Press: Lewes, Sussex.

Martin, N. G., Eaves, L. J., Heath, A. C., Jardine, R., Feindgold, L. M., & Eysenck, H. J. (1986). Transmission of social attitudes. Proceedings of the National Academy of Science, 83, 4364-4368.

Examples

data(twinData)
str(twinData)
plot(wt1 ~ wt2, data = twinData)
selVars = c("bmi1", "bmi2")
mzData <- subset(twinData, zyg == 1, selVars)
dzData <- subset(twinData, zyg == 3, selVars)

# equivalently
mzData <- subset(twinData, zygosity == "MZFF", selVars)

# Disregard sex, pick older cohort
mz <- subset(twinData, zygosity %in% c("MZFF","MZMM") & cohort == "older", selVars)

[Package OpenMx version 2.9.9 Index]