vech {OpenMx}	R Documentation

Half-vectorization

Description

This function returns the half-vectorization of an input matrix as a column vector.

Usage

vech(x)

Arguments

x
an input matrix.

Details

The half-vectorization of an input matrix consists of the elements in the lower triangle of the matrix, including the elements along the diagonal of the matrix, as a column vector. The column vector is created by traversing the matrix in column-major order.

See Also

vech2full, vechs, rvectorize, cvectorize

Examples

vech(matrix(1:9, 3, 3))
vech(matrix(1:12, 3, 4))

[Package OpenMx version 2.10.0 Index]