omxLocateParameters {OpenMx}	R Documentation

Get the location (model, matrix, row, column) and other info for a parameter

Description

Returns a data.frame summarizing the free parameters in a model, possibly filtered using ‘labels’.

For each located parameter, the label, model, matrix, row, col, value, and lbound & ubound are given as a row in the dataframe.

Duplicated labels return a row for each location in which they are found.

Usage

omxLocateParameters(model, labels = NULL, indep = FALSE)

Arguments

model
a MxModel object

labels
optionally specify which free parameters to retrieve.

indep
fetch parameters from independent submodels.

Details

Invoking the function with the default value for the ‘labels’ argument retrieves all the free parameters. The ‘labels’ argument can be used to select a subset of the free parameters. Note that ‘NA’ is a valid input to ‘labels’.

See Also

omxGetParameters, omxSetParameters, omxAssignFirstParameters

Examples

A <- mxMatrix('Full', 2, 2, labels = c("A11", "A12", NA, "A11"), values= 1:4,
   free = TRUE, byrow = TRUE, name = 'A')

model <- mxModel(A, name = 'model')

# Request all free parameters in model
omxLocateParameters(model)

# Request free parameters "A11" and all NAs
omxLocateParameters(model, c("A11", NA))

# Works with submodel
B = mxMatrix(name = 'B', 'Full', 1, 2, labels = c("B11", "notme"),
  free = c(TRUE, FALSE), values= pi)
model <- mxModel(model, mxModel(B, name = 'subB'))

# nb: only returns free parameters ('notme' not shown)
omxLocateParameters(model)

[Package OpenMx version 2.12.1 Index]